+++ src/test/java/com/revature/aop/ProjectAspectsTestSuite.java
@@ -1,4 +1,4 @@
-package com.revature.tests;
+package com.revature.aop;
 
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
